add strtobool function in utils
Python 3.12 removed distutils and this utility function is not included
in the setuptools package.
PEP 632[1] suggests that strtobool should be reimplemented rather than
relying on a compatibility package (unlike the rest of distutils).

This reimplementation doesn't accept all the inputs that distutils used
to accept but it is should work fine for our usage.

1: https://peps.python.org/pep-0632/